% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DALYfunc_individual.R
\name{f_DALY}
\alias{f_DALY}
\title{DALY function, individual}
\usage{
f_DALY(K, C = 0.16243, r, beta, a_death, a_disability, YLL_L, D)
}
\arguments{
\item{K}{Age weighting modulation factor (1=use age weighting, 0=no age weighting)}

\item{C}{contant (default = 0.16243)}

\item{r}{discount rate (between 0-1)}

\item{beta}{parameter of age weighting function (between 0-1)}

\item{a_death}{Average age of premature death due to disease for population (in years)}

\item{a_disability}{Average age of disease onset for population (in years)}

\item{YLL_L}{Life expectancy at age of death (in years)}

\item{D}{Disability weight (between 0-1)}
}
\description{
This function allows you to calculate Disability-Adjusted Life Years for an individual
}
\examples{
f_DALY()
}
\keyword{DALY}
