% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DALYfunc_population.R
\name{f_DALYpop}
\alias{f_DALYpop}
\title{DALY function, population}
\usage{
f_DALYpop(K, C = 0.16243, r, beta, a_death, a_disability, YLL_L, D,
  incident_deaths, incident_cases)
}
\arguments{
\item{K}{Age weighting modulation factor (1=use age weighting, 0=no age weighting)}

\item{C}{contant (default = 0.16243)}

\item{r}{discount rate (between 0-1)}

\item{beta}{parameter of age weighting function (between 0-1)}

\item{a_death}{Average age of premature death due to disease for population (in years)}

\item{a_disability}{Average age of disease onset for population (in years)}

\item{YLL_L}{Life expectancy at age of death (in years)}

\item{D}{Disability weight (between 0-1)}

\item{incident_deaths}{average number of deaths for the population}

\item{incident_cases}{average number of cases in the population}
}
\description{
This function allows you to calculate Disability-Adjusted Life Years for a population
}
\examples{
f_DALYpop()
}
\keyword{DALY}
